+++ processor/queuedprocessor/queued_processor_test.go
@@ -318,7 +318,7 @@ func TestTraceQueueProcessorHappyPath(t *testing.T) {
 	mockP.checkNumBatches(t, wantBatches)
 	mockP.checkNumSpans(t, wantSpans)
 
-	droppedView, err := findViewNamed(views, processor.StatDroppedSpanCount.Name())
+	droppedView, err := findViewNamed(views, "processor/"+processor.StatDroppedSpanCount.Name())
 	require.NoError(t, err)
 
 	data, err := view.RetrieveData(droppedView.Name)
@@ -326,7 +326,7 @@ func TestTraceQueueProcessorHappyPath(t *testing.T) {
 	require.Len(t, data, 1)
 	assert.Equal(t, 0.0, data[0].Data.(*view.SumData).Value)
 
-	data, err = view.RetrieveData(processor.StatTraceBatchesDroppedCount.Name())
+	data, err = view.RetrieveData("processor/" + processor.StatTraceBatchesDroppedCount.Name())
 	require.NoError(t, err)
 	assert.Equal(t, 0.0, data[0].Data.(*view.SumData).Value)
 	obsreporttest.CheckProcessorTracesViews(t, cfg.Name(), int64(wantSpans), 0, 0)