+++ receiver/receiverhelper/scrapercontroller_test.go
@@ -23,7 +23,6 @@ import (
 
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
-	"go.opencensus.io/stats/view"
 	"go.opencensus.io/trace"
 
 	"go.opentelemetry.io/collector/component/componenterror"
@@ -33,7 +32,6 @@ import (
 	"go.opentelemetry.io/collector/consumer/consumertest"
 	"go.opentelemetry.io/collector/consumer/pdata"
 	"go.opentelemetry.io/collector/exporter/exportertest"
-	"go.opentelemetry.io/collector/obsreport"
 	"go.opentelemetry.io/collector/obsreport/obsreporttest"
 )
 
@@ -181,8 +179,6 @@ func TestScrapeController(t *testing.T) {
 		},
 	}
 
-	views := obsreport.Configure(false, true)
-
 	for _, test := range testCases {
 		t.Run(test.name, func(t *testing.T) {
 			trace.ApplyConfig(trace.Config{DefaultSampler: trace.AlwaysSample()})
@@ -191,8 +187,9 @@ func TestScrapeController(t *testing.T) {
 			trace.RegisterExporter(ss)
 			defer trace.UnregisterExporter(ss)
 
-			require.NoError(t, view.Register(views...))
-			defer view.Unregister(views...)
+			done, err := obsreporttest.SetupRecordedMetricsTest()
+			require.NoError(t, err)
+			defer done()
 
 			initializeChs := make([]chan bool, test.scrapers+test.resourceScrapers)
 			scrapeMetricsChs := make([]chan int, test.scrapers)
@@ -358,7 +355,12 @@ func assertReceiverSpan(t *testing.T, spans []*trace.SpanData) {
 }
 
 func assertReceiverViews(t *testing.T, sink *exportertest.SinkMetricsExporter) {
-	obsreporttest.CheckReceiverMetricsViews(t, "receiver", "", int64(sink.MetricsCount()), 0)
+	dataPointCount := 0
+	for _, md := range sink.AllMetrics() {
+		_, dpc := md.MetricAndDataPointCount()
+		dataPointCount += dpc
+	}
+	obsreporttest.CheckReceiverMetricsViews(t, "receiver", "", int64(dataPointCount), 0)
 }
 
 func assertScraperSpan(t *testing.T, expectedErr error, spans []*trace.SpanData) {
@@ -399,6 +401,9 @@ func assertScraperViews(t *testing.T, expectedErr error, sink *exportertest.Sink
 func singleMetric() pdata.MetricSlice {
 	metrics := pdata.NewMetricSlice()
 	metrics.Resize(1)
+	metrics.At(0).SetDataType(pdata.MetricDataTypeIntGauge)
+	metrics.At(0).IntGauge().InitEmpty()
+	metrics.At(0).IntGauge().DataPoints().Resize(1)
 	return metrics
 }